pyproject.toml
[project]
name = "linguaflow-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.0",
    "gunicorn>=23.0.0",
    "langchain>=1.1.3",
    "langchain-core>=1.1.3",
    "langchain-groq>=1.1.0",
    "mlflow>=3.7.0",
    "psycopg2>=2.9.11",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "uvicorn>=0.38.0",
]
